Simplify AbstractActivityCollector API
The previous API required subclasses to provide a pair of methods
(#raw_activities and #build_activity_from_raw_data).  This requirement
dictates too many implementation details.  Subclasses need to provide a
collection of activities to be saved, and that's it.  The API shouldn't
care *how* the subclasses construct the activities.

  - Refactor the API to only require an #activities method.
  - Refactor subclasses to support the updated API.

Thanks to @jdpace for the suggestion.