Use the episode number field when present for iTunes TV shows
For a given TV show in an iTunes Music Library XML file, the "Track Number"
and the "Episode Order" field usually have the same value, but they can
sometimes differ.  If the "Episode Order" field is present, use its value
as the episode number.  Only use the "Track Number" field as the episode
number if the "Episode Order" field is not present.