Replace iTunes parsing library
Swap out "itunes_parser" for "itunes-library".

In order to extract each of the collectors out into a gem, those
gems need to be able to declare dependencies on other gems.  (For
example, stratify-itunes will need to declare a dependency on the
iTunes parsing gem that it uses.)  itunes_parser is not published
as a gem; itunes-library *is* published as a gem, so it meets this
need.

This commit declares a dependency on a branch of itunes-library
that includes some additional functionality that Stratify needs. I
suspect (and hope) that we'll see a new version of the itunes-library
gem published soon, incorporating the changes in this branch.  At
that point, we can declare the dependency on the new version of the
gem (instead of depending on a branch on GitHub).