Use thoughtafter-simple-rss fork for RSS processing
In order to extract each of the collectors out into a gem, those
gems need to be able to declare dependencies on other gems.  (For
example, stratify-instapaper will need to declare a dependency on
the RSS parsing gem that it uses.)

The thoughtafter fork of simple-rss seems to meet Stratify's needs,
and it is also published as a gem, so let's use it.