Fix encoding bug seen when reading remote iTunes XML files
Prior to this commit, if you attempted to run the iTunes collector reading
from an iTunes library XML file living on a remote host (e.g., via HTTP, FTP),
it failed with the following error:

> Error running collector.  Collector => #<Stratify::ITunes::Collector:0x007f8c70f59340>.
> Error => incompatible encoding regexp match (UTF-8 regexp with ASCII-8BIT string)"

This commit:

  - Adds a spec that recreates the encoding-related bug and asserts the
    ability to collect data from an iTunes XML file over HTTP.
  - Adds VCR and fakeweb for mocking out the network access during the
    specs.
  - Adds a VCR cassette (i.e., fixture). Note that the VCR cassette
    was created using ths same iTunes XML file that resides at
    stratify-itunes/spec/fixtures/iTunes Music Library.xml.
  - Forces the encoding to UTF-8 after reading the iTunes XML file
    (which allows the spec to pass, and fixes the bug, as far as best
    can tell).

Some thoughts on the fix: iTunes XML files use UTF-8 encoding, as advertised
at the top of any iTunes XML file ...

    <?xml version="1.0" encoding="UTF-8"?>

Given this convention, it feels safe to force the encoding to UTF-8.  I have
not yet encountered any situations where this is problematic, but I admit to
not be _fully_ confident in this solution.

Hat tip to James Edward Gray II for his excellent articles on Ruby 1.9
encoding, of which this one in particular helped me track down a solution to
this bug:

http://blog.grayproductions.net/articles/ruby_19s_string