Ignore iTunes tracks with no 'Play Date UTC' value
In the "iTunes Music Library.xml" file, it is possible for a track to
have a "Play Count" value but no "Play Date UTC" value. Essentially,
iTunes is reporting that the track has been played n times, but iTunes
does not know *when* it was last played.

Prior to this commit, Stratify looked for all "played" tracks (using
the #played? method in the itunes-library gem) and then sorted them by
"Play Date UTC". The #played method looks at the "Play Count" value,
so it was possible to get a track that is considered to be played, but
has no "Play Date UTC". Since Stratify uses the "Play Date UTC" to
determine the activity timestamp (i.e., when did this activity occur), a
null value is problematic.

This reveals the fact that we really don't care whether iTunes thinks a
track has a "Play Count". Instead, we really only care whether iTunes
can tell us when the playing actually occurred. In other words, we
should just ignore any tracks that don't have a "Play Date UTC".

So ...

  - Update Stratify::ITunes::Query to only select tracks that have a
    "Play Date UTC" value (i.e., a non-null last_played_at value).
  - Update the test fixtures to include an example of a track that has a
    "Play Count" but no "Play Date UTC" value, and add a spec for this
    scenario.