Move stratify-base into Rails app (WIP)
This commit moves stratify-base into the Rails app, but there are a few
failing tests. It looks like we'll have to make some minor code changes
to allow the **full** suite of Stratify tests to run together.

The stratify-base test suite affects the content of
Collector.collector_classes, and that causes problems for tests
outside of the stratify-base test suite. One test in the test suite
should not have side effects observable by other tests in the test
suite. So, we've got a little work to do in order to remove the current
side effects.