Require explicit registration of collector classes
As noted in ada607e, we had a few failing tests due to the stratify-base
test suite manipulating the content of Collector.collector_classes. This
commit fixes those failing tests.

Instead of using the .inherited hook to implicitly register collector
classes when a subclass of Stratify::Collector is defined, we now
require explicit registration of collector classes. By removing the
dependency on the inherited hook, we can register the collector
classes outside of the class definition lifecycle. We can clear the
registered classes whenever we like, and we can re-register them as
needed (which is handy for testing). This also means that we no longer
*require* that a collector class inherit from Stratify::Collector.